# Conflicts:
#	Songbird/machine_learning_and_dsp/Test_noise_reduce.py
#	Songbird/machine_learning_and_dsp/TestandTrain.py
#	Songbird/machine_learning_and_dsp/test_model.py

+++ Songbird/machine_learning_and_dsp/tbl_create.sql
@@ -4,7 +4,7 @@ CREATE TABLE sampleInfo (
   added     TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   latitude  FLOAT,
   longitude FLOAT,
-  humidity  INTEGER,
+  humidity  FLOAT,
   temp      FLOAT,
   light     FLOAT,
   type1     VARCHAR(40),
